\name{modreg-package}

\alias{modreg-package}

\docType{package}

\title{ modreg }

\description{
    This package provides a single function for exploring interactions in moderated multiple regression: \emph{modreg}. \emph{modreg}, is used to conduct moderated multiple regression and simple slope analyses and provide simple slope plots for a single predictor and one or two continuous or categorical moderators.
}
