% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ModReg}
\alias{ModReg}
\title{Modulation mode inference}
\usage{
ModReg(
  exprMatrix,
  modulator,
  regulator,
  regulon,
  cutoff.modulator = 1/4,
  permut_num = 1000,
  mc.cores = 1
)
}
\arguments{
\item{exprMatrix}{Gene expression matrix with gene symbol (row) x sample (column).}

\item{modulator}{A gene symbol for the modulator.}

\item{regulator}{A gene symbol for the regulator}

\item{regulon}{A vector for the target genes of the regulator.}

\item{cutoff.modulator}{A number (1/5 <= cutoff <= 1/2) defining the samples with highly and lowly expressed modulator.}

\item{permut_num}{Permutation number for differential mutual information analysis.}

\item{mc.cores}{Number of CPU cores.}
}
\value{
An ModReg object
}
\description{
Calculate the modulation relation between a modulator and a regulator.
}
\details{
ModReg is comprised of two steps. At first, all tripartite relations of Modulator-Regulator-Target were estimated based on differential connectivity analysis. Specifically, all tumor samples were separated into the top (M+) and bottom (M−) of samples in which the modulator was most and least expressed. In these two sets (M+ and M−), the significance of differential correlation relationships between one regulator and its targets were quantified using the difference in z-scores by Fisher z-transformation. A two-sided p value can be calculated using the standard normal distribution, and Benjamini–Hochberg method was used to adjust P-values for multiple hypothesis tests. Significant tripartite relations were kept for the next step (adjusted p<0.05).

Secondly, ModReg calculated the differential average mutual information (DAMI) of the significant regulated targets from step 1, which equals to the difference of average mutual information (i.e., dependence relationship) of all significant regulated targets between M+ and M− sample sets. Thus, DAMI quantified how much the modulator was able to affect the fucntion of the regulator (DAMI > 0 for activation and DAMI < 0 for inhibition, respectively). Permutation test is used to estimate the empirical distribution of DAMI by randomly selecting a target set of equal size with the significant regulated targets from the complete regulon of the regulator, and the associated p value was estimated as the percentage of random trials with an absolute value of DAMI greater than the absolute value of the measured DAMI. Based on the DAMI and p value here, users can determine the modulation mode of the modulator and the regulator.
}
\examples{
load(file.path(system.file(package = "ModReg"), "extdata/exampleData.rda"))
ModReg_obj <- ModReg(exprMatrix, modulator="SIRT7", regulator="NFE2L2", regulon)

}
