% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotModReg}
\alias{plotModReg}
\title{Modulator-Regulator visualization}
\usage{
plotModReg(exprMatrix, ModReg_obj)
}
\arguments{
\item{exprMatrix}{Gene expression matrix with gene symbol (row) x sample (column).}

\item{ModReg_obj}{An ModReg object}
}
\value{
A ggplot2 object
}
\description{
Visualize the modulation relation between a modulator and a regulator.
}
\details{
The left and right panels present the patient group with low (M−) and high (M+) level of SIRT7 expression, in which rows are targets and columns are patient samples, respectively. In each patient group (M− or M+), the samples are sorted based on NFE2L2 expression level.
}
\examples{
load(file.path(system.file(package = "ModReg"), "extdata/exampleData.rda"))
ModReg_obj <- ModReg(exprMatrix, modulator="SIRT7", regulator="NFE2L2", regulon)
plotModReg(exprMatrix, ModReg_obj)

}
