% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotModRegTar}
\alias{plotModRegTar}
\title{Modulator-Regulator-Target visualization}
\usage{
plotModRegTar(exprMatrix, modulator, regulator, target, cutoff.modulator = 1/4)
}
\arguments{
\item{exprMatrix}{Gene expression matrix with gene symbol (row) x sample (column).}

\item{modulator}{A gene symbol for the modulator.}

\item{regulator}{A gene symbol for the regulator.}

\item{target}{A target gene of the regulator.}

\item{cutoff.modulator}{A number (1/5 <= cutoff <= 1/2) defining the samples with highly and lowly expressed modulator.}
}
\value{
A ggplot2 object
}
\description{
ModReg
}
\examples{
load(file.path(system.file(package = "ModReg"), "extdata/exampleData.rda"))
plotModRegTar(exprMatrix, modulator="SIRT7", regulator="NFE2L2", target="IRF2")

}
