\name{SimpleRegression-package}

\alias{SimpleRegression-package}

\docType{package}

\title{ SimpleRegression }

\description{
    This package provides functions for exploring interactions in moderated multiple regression: \emph{mod_regression} and \emph{plot_mr}. The first function, \emph{mod_regression}, is used to conduct moderated multiple regression and simple slope analyses for a single predictor and one or two continuous or categorical moderators. The second function, \emph{plot_mr}, provides simple slope plots.
}
