% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{ModReg}
\alias{ModReg}
\title{Calculate the modulation relation between a modulator and a regulator}
\usage{
ModReg(exprMatrix, modulator, regulator, regulon, mc.cores=1)
}
\arguments{
\item{exprMatrix}{a numeric data frame containing data from expression profile where columns contain samples and rows contain genes}

\item{modulator}{a character variable which define the modulator gene}

\item{regulator}{a character variable which define the regulator gene}

\item{regulon}{a character vector with target genes of regulator}

\item{mc.cores}{number of cpu cores using in parallel computation}
}
\value{
Typically, the returned list contains two objects. The first one is a data frame of the table of differential correlations between low (CS) and high (TS) expression of modulator. The second one is matirx showing relation between the modulator and regulator, percentage of influential targets of the regulator.
}
\description{
ModReg is a function that evaluate the differential connectivity of a regulator and its targets between high and low expression level of a modulator to identify the modulation relation
}
\examples{
load(file.path(system.file(package = "ModReg"), "extdata/exampleData.rda"))
res <- ModReg(exprMatrix, modulator, regulator, regulon)
}
